Adding Games to Collection \Users
addGame() now working in shipGame.js , however, having trouble calling it in the right place.

Also, leaderboards will now pull directly from \Users collection. Decided it was redundant to create another collection.

Lastly, It looks like users are not being populated into the database when they create an account. It's not until they login that they'll be added to the database.